+++ ReleaseNotes.md
@@ -33,6 +33,7 @@
 - [Technical] - Speedup the Xiaomi pairing/provisioing process
 - [Technical] - Able to set PowerOn after a powerOff ( 0 for Off, 1 for On, 255, for previous state), Works for devices able to handle it. (Legrand, Philips)
 - [Technical] - Friendly Provisionning
+- [Technical] - Management of Zigate Blue Led. No need to restart plugin. Action taken immediatly from settings
 
 ##             - 4.5.0 ( Web User Interface ) https://github.com/sasu-drooz/Domoticz-Zigate/releases/tag/4.5.0
 - [Technical] - Web Server Class